Move helper methods to a common header
